align zte tpd int void kcfi family

+++ kernel_development/drivers/reconstructed/zte_tpd/suspend_tp_need_awake.c
@@ -1,10 +1,10 @@
 #include "defs.h"
 
-__int64 suspend_tp_need_awake(void)
+int suspend_tp_need_awake(void)
 {
-  __int64 (*v0)(void); // x8
+  int (*v0)(void); // x8
 
-  v0 = *(__int64 (**)(void))(tpd_cdev + 3248);
+  v0 = *(int (**)(void))(tpd_cdev + 3248);
   if ( !v0 )
     return 0;
   if ( *((_DWORD *)v0 - 1) != -2139014615 )
